web_content_agent/requirements.txt
# 核心依赖
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
python-dotenv>=1.0.0

# HTML解析和处理
lxml>=4.9.0
html5lib>=1.1

# HTTP请求增强
urllib3>=1.26.0

# 日志和工具
colorama>=0.4.6
click>=8.1.0

# 开发和测试依赖（可选）
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# 性能优化（可选）
cchardet>=2.1.7